Move file archive modes to separate folder (#2205)
