+++ src/NLog/NLog.wp8.csproj
@@ -416,14 +416,14 @@
     <Compile Include="Targets\DefaultJsonSerializer.cs" />
     <Compile Include="Targets\EventLogTarget.cs" />
     <Compile Include="Targets\EventLogTargetOverflowAction.cs" />
-    <Compile Include="Targets\FileArchiveModeBase.cs" />
-    <Compile Include="Targets\FileArchiveModeDate.cs" />
-    <Compile Include="Targets\FileArchiveModeDateAndSequence.cs" />
-    <Compile Include="Targets\FileArchiveModeDynamicSequence.cs" />
-    <Compile Include="Targets\FileArchiveModeDynamicTemplate.cs" />
-    <Compile Include="Targets\FileArchiveModeFactory.cs" />
-    <Compile Include="Targets\FileArchiveModeRolling.cs" />
-    <Compile Include="Targets\FileArchiveModeSequence.cs" />
+    <Compile Include="Targets\FileArchiveModes\FileArchiveModeBase.cs" />
+    <Compile Include="Targets\FileArchiveModes\FileArchiveModeDate.cs" />
+    <Compile Include="Targets\FileArchiveModes\FileArchiveModeDateAndSequence.cs" />
+    <Compile Include="Targets\FileArchiveModes\FileArchiveModeDynamicSequence.cs" />
+    <Compile Include="Targets\FileArchiveModes\FileArchiveModeDynamicTemplate.cs" />
+    <Compile Include="Targets\FileArchiveModes\FileArchiveModeFactory.cs" />
+    <Compile Include="Targets\FileArchiveModes\FileArchiveModeRolling.cs" />
+    <Compile Include="Targets\FileArchiveModes\FileArchiveModeSequence.cs" />
     <Compile Include="Targets\FileArchivePeriod.cs" />
     <Compile Include="Targets\FilePathKind.cs" />
     <Compile Include="Targets\FileTarget.cs" />
