+++ src/NLog/Targets/FileArchiveModes/FileArchiveModeDateAndSequence.cs
@@ -31,13 +31,13 @@
 // THE POSSIBILITY OF SUCH DAMAGE.
 // 
 
-namespace NLog.Targets
-{
-    using System;
-    using System.Collections.Generic;
-    using System.Globalization;
-    using System.IO;
+using System;
+using System.Collections.Generic;
+using System.Globalization;
+using System.IO;
 
+namespace NLog.Targets.FileArchiveModes
+{
     /// <summary>
     /// <para>
     /// Archives the log-files using a date and sequence style numbering. Archives will be stamped
